Add single and multi wildcard options to all tables
Since ibus-table 1.8.3, the default for single and multi wildcards is
empty (Because the most natural way to switch off wildcards seems to
be to make the value of the wildcard options the empty string).

Therefore, for the tables where wildcard support is useful, wildcard
options should be added to have wildcards switched on by default.