Change MAX_KEY_LENGTH from 2 to 1 for the rusle, rustrad, thai, and yawerty tables.
The comments why MAX_KEY_LENGTH was set to 2 said:

    “changed to 2 to fake col prob in sqlite db”

    “It is supposed to be 1 but need to be 2 to avoid schema mismatch
    bug in optimize_database() in ibus-table.”

So this was probably a workaround for some bug in ibus-table,
probably it had something to do with the old database format
where each input key was stored in a different column.

With the current ibus-table dataformat, there is no problem
in using the correct value MAX_KEY_LENGTH=1.