Use better localized names for the rusle table
Because of LANGUAGES = ru (or LANGUAGES = ru_RU), the (possibly
translated) language name is added as a prefix to the (possibly
localized) long name of the input method anyway when displayed by in
the Gnome control centre or in the input method selection menu of
gnome. When listing with “ibus list-engine” it is in the section for
the language.

Therefore, adding the language again as a prefix to the input method
name or the localized input method names is superfluous.

Using the language only in localized tables names, i.e.

    NAME.ru = устаревшая

instead of

    NAME.ru_RU = устаревшая

requires ibus-table >= 1.9.1.